requirements.txt
# requirements.txt — Cài đặt: pip install -r requirements.txt

# HTTP & feeds
requests
feedparser

# Tài chính & dữ liệu
yfinance
pandas
numpy

# Tạo file Word & PDF
python-docx
docx2pdf        # Cần MS Word (Win/macOS) hoặc LibreOffice (Linux)

# Template
Jinja2

# Cấu hình môi trường
python-dotenv

# Optional — PostgreSQL
# psycopg2-binary==2.9.9

# Optional — Redis
# redis==5.0.1
